requirements.txt
h11==0.8.1
jaconv==0.2.4
Janome==0.3.9
Jinja2==2.10.1
PyYAML==5.1.2
regex==2019.8.19
requests==2.22.0
responder==1.3.2